store and propagate raw helix language config
This is in preparation for supporting arbitrary file path globs.
An arbitrary glob will produce a synthetic language just for that glob.
In order to avoid losing most of the language configuration, we will
attempt to detect the language the glob matches against and copy the
entire config from the original language to the synthetic one.

This also includes a migration from toml to toml_edit. This will make it
possible to produce synthetic language configurations that look almost
identical to the original ones.